requirements.txt
crewai==0.86.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
langchain>=0.1.0
pydantic>=2.5.2
cryptography>=41.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
